==> R/USR/06 Conc Realizations-Data Setup.R <==
##### General Setup #####
source("~/Projects/Calcs/R/USR/Pre-Loader.R")

setwd(modelsFile)

load("qin.Rdata")
load("ecin.Rdata")
load("tin.Rdata")
load("qout.Rdata")
load("ecout.Rdata")
load("tout.Rdata")
load("qCAN.Rdata")
load("qCON.Rdata")
load("qFLS.Rdata")
load("qFLY.Rdata")
load("qHOL.Rdata")
load("qHRC.Rdata")
load("qRFD.Rdata")
load("qRFR.Rdata")
load("qTIM.Rdata")
load("qWTP.Rdata")

q.in <- as.vector(as.matrix(q.in))
ec.in <- as.vector(as.matrix(ec.in))
t.in <- as.vector(as.matrix(t.in))
q.out <- as.vector(as.matrix(q.out))
ec.out <- as.vector(as.matrix(ec.out))
t.out <- as.vector(as.matrix(t.out))
q.CAN <- as.vector(as.matrix(q.CAN))
q.CON <- as.vector(as.matrix(q.CON))
q.FLS <- as.vector(as.matrix(q.FLS))
q.FLY <- as.vector(as.matrix(q.FLY))
q.HOL <- as.vector(as.matrix(q.HOL))
q.HRC <- as.vector(as.matrix(q.HRC))
q.RFD <- as.vector(as.matrix(q.RFD))
q.RFR <- as.vector(as.matrix(q.RFR))
q.TIM <- as.vector(as.matrix(q.TIM))
q.WTP <- as.vector(as.matrix(q.WTP))

calcData <-data.frame(q.in, ec.in, t.in, q.out, ec.out, t.out, q.CAN, q.CON, q.FLS, q.FLY, q.HOL, q.HRC, q.RFD, q.RFR, q.TIM, q.WTP)
save(calcData, file="calcData.Rdata")

rm(list=ls())
